Apagar Formas antigas de requisição
