--- app/src/main/java/com/example/geoshot/generalUtilities/AuxiliarGeral.java
@@ -1,9 +1,0 @@
-package com.example.geoshot.generalUtilities;
-
-public class AuxiliarGeral {
-    private static final String IP_SERVER_ADDRESS = "172.15.2.239";
-
-    public static String getIPServerAddress() {
-        return IP_SERVER_ADDRESS;
-    }
-}
