+++ app/src/main/java/com/example/geoshot/generalUtilities/post/APIClientPost.java
@@ -1,7 +1,6 @@
-package com.example.geoshot.generalUtilities.post0;
+package com.example.geoshot.generalUtilities.post;
 
 import com.example.geoshot.generalUtilities.get.APIClient0;
-import com.example.geoshot.generalUtilities.get.APIClientGet;
 
 public class APIClientPost extends APIClient0 {
