+++ app/src/main/java/com/example/geoshot/generalUtilities/post/PostSignup.java
@@ -1,4 +1,4 @@
-package com.example.geoshot.generalUtilities.post0;
+package com.example.geoshot.generalUtilities.post;
 
 import android.util.Log;
