--- app/src/main/java/com/example/geoshot/generalUtilities/post/PostStrategy.java
@@ -1,6 +1,0 @@
-package com.example.geoshot.generalUtilities.post;
-
-public interface PostStrategy {
-    public String json();
-    public String url();
-}
